Move snake case handling to the right place
